# site/openapi.yaml
openapi: 3.0.0
info:
  title: openapi
  version: '1.0'
  description: An abstract API that can be used to send messages.
  contact:
    name: FooBar
servers:
  - url: 'http://localhost:8000'
    description: Localhost
paths:
  /sendMessage:
    post:
      summary: Send Message
      operationId: post-sendMessage
      responses:
        '202':
          description: The message request was accepted and will be processed by the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          headers:
            X-Request-ID:
              schema:
                type: string
              description: Unique ID identifying this request for debugging purposes
        '400':
          description: The request was rejected due to invalid input. See the response for reasons.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
      description: Send a message via the defined channel
      tags:
        - messaging
    parameters: []
  /uploadMedia:
    post:
      summary: Upload Media File
      operationId: post-uploadMedia
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties: {}
        description: The media files that you want to upload.
      tags:
        - media
      description: This endpoint can be used to upload media files.
  '/downloadMedia/{media_id}':
    get:
      summary: Download Media File
      tags:
        - media
      responses:
        '200':
          description: Media has been downloaded
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
      operationId: get-downloadMedia
      description: This endpoint can be used to download media files.
    parameters:
      - schema:
          type: string
        name: media_id
        in: path
        required: true
        description: The ID of the media file
components:
  schemas:
    APIResponse:
      title: APIResponse
      type: object
      properties:
        status:
          type: string
          enum:
            - success
            - failed
        reason:
          type: string
        errors:
          type: object
        id:
          type: string
      required:
        - status
    TextPayload:
      title: TextPayload
      type: object
      properties:
        type:
          type: string
          enum:
            - text
        body:
          type: string
      required:
        - body
    MediaPayload:
      title: MediaPayload
      type: object
      properties:
        type:
          type: string
          enum:
            - image
            - video
            - audio
            - document
        url:
          type: string
        text:
          type: string
      required:
        - type
        - url
    MessageRequest:
      title: MessageRequest
      type: object
      properties:
        to:
          type: string
        type:
          type: string
          enum:
            - text
            - media
        provider:
          type: string
        payload:
          type: array
          minItems: 1
          uniqueItems: false
          items:
            oneOf:
              - $ref: '#/components/schemas/TextPayload'
              - $ref: '#/components/schemas/MediaPayload'
              - $ref: '#/components/schemas/ParameterPayload'
      required:
        - to
        - type
        - provider
        - payload
    ParameterPayload:
      title: ParameterPayload
      type: object
      properties:
        type:
          type: string
          enum:
            - parameter
        payload:
          $ref: ''
        '':
          type: string
  securitySchemes: {}
tags:
  - name: media
  - name: messaging